refactor: make steps order step components
