+++ src/views/refund/steps/index.js
@@ -1,0 +1,5 @@
+import InputDestintionAddress from './inputDestintionAddress';
+import CompleteRefund from './completeRefund';
+import UploadRefundFile from './uploadRefundFile';
+
+export { InputDestintionAddress, UploadRefundFile, CompleteRefund };
