+++ src/views/reverse/steps/index.js
@@ -1,0 +1,4 @@
+import PayInvoice from './payInvoice';
+import InputAddress from './inputAddress';
+
+export { PayInvoice, InputAddress };
