+++ src/views/swap/steps/index.js
@@ -1,0 +1,6 @@
+import InputInvoice from './inputInvoice';
+import SendTransaction from './sendTransaction';
+import DownloadRefund from './downloadRefund';
+import CompleteSwap from './completeSwap';
+
+export { InputInvoice, SendTransaction, DownloadRefund, CompleteSwap };
